[IMP] website: mega menu and dropdowns on mobile The current mobile navigation header is unusable when there are submenus or a mega_menu. This commit improves the usability by implementing a "fake" offcanvas on mega_menu on mobile (essentially a dropdown which behaves visually as an offcanvas).
For the submenus, we transform the dropdowns into accordions to improve
the navigation experience.

Because of the new headers we also have to apply this behavior on
hamburger and sidebar header, thus we can't rely on breakpoints to
manage the styling since these headers are shown on desktop sized
screens.

task-3188239

Part-of: #140083
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>